Fix error occured when the compared date is invalid
For instance, 2018-09-24 is a holiday in Taiwan so there is no currency record on that date. However, an error would occur on 2018-10-01 when the app computes the currency change in last week because it can't find the corresponding value. This commit makes the app search for the last valid date for computation.